backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==12.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==1.5.3
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
scipy==1.11.4